# Conflicts:
#	src/main/java/com/example/project11/Controllers/SeparateQuestions.java

+++ src/main/java/com/example/project11/ProjectInfo/STEP2/GraduatePassingGradePredictionThreshold.java
@@ -21,22 +21,22 @@ public class GraduatePassingGradePredictionThreshold {
     public void calculatePassingPercentagesArray() {
         coursePassingMap = new HashMap<>();
         // Calculate passing percentages
-            double[] passingPercentages = calculatePassingPercentagesArray(allStudents);
+            int[] passingPercentages = calculatePassingPercentagesArray(allStudents);
 
             for (int i = 0; i < passingPercentages.length; i++) {
-                coursePassingMap.put(Integer.toString(i), (int)passingPercentages[i]); // Course numbers start at 1
+                coursePassingMap.put(Integer.toString(i), passingPercentages[i]); // Course numbers start at 1
             }
 
 
     }
     // a method for calculating a passing percentage
-    private double[] calculatePassingPercentagesArray(double[][] matrix) {
+    private int[] calculatePassingPercentagesArray(double[][] matrix) {
         // we have determined that a passing student is one that has a 6 or higher based on the graduating grades
-        double[]passingPercentages = new double[matrix[0].length];
+        int[]passingPercentages = new int[matrix[0].length];
 
         for( int j = 0 ; j< allStudents[0].length ; j++) {
-            double sum =  0 ;
-            double count = 0 ;
+            int sum =  0 ;
+            int count = 0 ;
             for( int i = 0 ; i< allStudents.length ; i++) {
                 count++;
                 if(allStudents[i][j]>= 6.0){
@@ -53,6 +53,8 @@ public class GraduatePassingGradePredictionThreshold {
     // a getter for the hash map
     public Map<String, Integer> getCoursePassingMap() {
         calculatePassingPercentagesArray();
+
+
         return coursePassingMap;
     }
 }
